@@ -72,7 +72,7 @@
 	</scm>
 
 	<properties>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 		<maven.compiler.release>${java.version}</maven.compiler.release>
@@ -85,7 +85,7 @@
 		<!-- Dependencies -->
 		<guice.version>5.1.0</guice.version>
 		<junit.version>1.11.3</junit.version>
-		<xtext.version>2.37.0</xtext.version>
+		<xtext.version>2.40.0</xtext.version>
 		<jacoco.version>0.8.12</jacoco.version>
 		<vaadin.version>24.8.8</vaadin.version>
 		<okhttp.version>4.12.0</okhttp.version>
@@ -103,7 +103,7 @@
 		<json-schema-validator.version>1.5.4</json-schema-validator.version>
 
 		<!-- Maven Plug-Ins -->
-		<tycho.version>4.0.10</tycho.version>
+		<tycho.version>5.0.0</tycho.version>
 		<sb-contrib.version>7.6.8</sb-contrib.version>
 		<spdx-maven-plugin.version>1.0.3</spdx-maven-plugin.version>
 		<license-maven-plugin.version>4.6</license-maven-plugin.version>
@@ -374,7 +374,7 @@
 					<version>3.4.2</version>
 				</plugin>
 				<plugin>
-					<groupId>org.eclipse.xtend</groupId>
+					<groupId>org.eclipse.xtext</groupId>
 					<artifactId>xtend-maven-plugin</artifactId>
 					<version>${xtext.version}</version>
 					<executions>
@@ -742,8 +742,7 @@
 							<version>${maven-surefire-plugin.version}</version>
 							<configuration>
 								<argLine>@{argLine} -Xshare:off
-									-XX:+EnableDynamicAgentLoading
-									-Djava.security.manager=allow</argLine>
+									-XX:+EnableDynamicAgentLoading</argLine>
 							</configuration>
 						</plugin>
 						<plugin>
@@ -752,8 +751,7 @@
 							<version>${maven-failsafe-plugin.version}</version>
 							<configuration>
 								<argLine>@{argLine} -Xshare:off
-									-XX:+EnableDynamicAgentLoading
-									-Djava.security.manager=allow</argLine>
+									-XX:+EnableDynamicAgentLoading</argLine>
 							</configuration>
 						</plugin>
 					</plugins>
